+++ src/app/shared/components/grid-usage-link/grid-usage-link.component.scss
