+++ src/app/shared/components/grid-usage-link/grid-usage-link.component.ts
@@ -1,0 +1,38 @@
+import { Component, Input } from '@angular/core';
+import { MatDialog } from '@angular/material/dialog';
+import { GridUsagesDialogComponent } from '../grid-usages-dialog/grid-usages-dialog.component';
+import { RegistrationEntityTypes } from '../../models/registrations/registration-entity-categories.model';
+
+@Component({
+  selector: 'app-usage-link',
+  standalone: false,
+  templateUrl: './grid-usage-link.component.html',
+  styleUrl: './grid-usage-link.component.scss',
+})
+export class UsageLinkComponent {
+  constructor(private dialog: MatDialog) {}
+
+  @Input() usages!: string[];
+  @Input() name: string = '';
+  @Input() type: RegistrationEntityTypes | undefined;
+
+  onUsageClick(event: Event) {
+    console.log(this.type);
+    event.preventDefault();
+    this.dialog.open(GridUsagesDialogComponent, {
+      data: { usages: this.usages, title: this.getTitle() },
+    });
+  }
+
+  getTitle() {
+    return this.getPrefix() + this.name;
+  }
+
+  getPrefix() {
+    if (this.type === 'it-interface') {
+      return $localize`Organisationer der anvender snitfladen: `;
+    } else {
+      return $localize`Anveldelser af `;
+    }
+  }
+}
