+++ src/app/shared/components/grid-usages-dialog/grid-usages-dialog.component.ts
@@ -3,9 +3,9 @@ import { MAT_DIALOG_DATA } from '@angular/material/dialog';
 
 @Component({
   selector: 'app-usages',
-  templateUrl: './usages.component.html',
-  styleUrls: ['./usages.component.scss'],
+  templateUrl: './grid-usages-dialog.component.html',
+  styleUrls: ['./grid-usages-dialog.component.scss'],
 })
-export class UsagesComponent {
+export class GridUsagesDialogComponent {
   constructor(@Inject(MAT_DIALOG_DATA) public data: { usages: string[]; title: string }) {}
 }
