+++ src/app/shared/components/grid/grid.component.html
@@ -158,13 +158,7 @@
       </ng-template>
 
       <ng-template *ngIf="column.style === 'usages'" kendoGridCellTemplate let-dataItem>
-        <app-usage-link [isInterface]="false" [name]="searchProperty(dataItem, 'Name')" [usages]="searchProperty(dataItem, column.field)">
-
-        </app-usage-link>
-      </ng-template>
-
-      <ng-template *ngIf="column.style === 'usages-interface'" kendoGridCellTemplate let-dataItem>
-        <app-usage-link [isInterface]="true" [name]="searchProperty(dataItem, 'Name')" [usages]="searchProperty(dataItem, column.field)">
+        <app-usage-link [type]="column.entityType" [name]="searchProperty(dataItem, 'Name')" [usages]="searchProperty(dataItem, column.field)">
 
         </app-usage-link>
       </ng-template>
